perf(parser-adapter-json): use tree-sitter cursor for CST traversal
Refs #691
